molecule: default: apply the role twice to reproduce #44
https://github.com/theoapp/ansible-role-theo-agent/issues/44 exposed a bad
manipulation of the variable theo_agent_sshd_authorized_keys_command. This
commit extends default molecule scenario to apply the role twice: this is managed
by tricking ansible with a fake variable to force applying the role twice.

If we had no such bug as #44, this would produce no changes when the role is
re-applied, hence idempotence test is expected to succeed. This is not the case
due to #44.